# rlp, trie: faster trie node encoding (#24126)