--- a/trie/committer.go
+++ b/trie/committer.go
@@ -35,7 +35,6 @@ const leafChanSize = 200
 // capture all dirty nodes during the commit process and keep them cached in
 // insertion order.
 type committer struct {
-	tmp sliceBuffer
 	sha crypto.KeccakState
 
 	owner       common.Hash // TODO: same as nodes.owner, consider removing
@@ -48,7 +47,6 @@ type committer struct {
 var committerPool = sync.Pool{
 	New: func() interface{} {
 		return &committer{
-			tmp: make(sliceBuffer, 0, 550), // cap is as large as a full fullNode.
 			sha: sha3.NewLegacyKeccak256().(crypto.KeccakState),
 		}
 	},