@huyngopt1994 2024-10-29
This PR is LGTM, Could u add description related to some minor refactors in rlp package and help to resolve conflicts before merging too ?